refactor(surveys): thin SurveyContainer down for S3776 [ENG-2334]
SonarQube flagged SurveyContainer at cognitive complexity 16 against a limit
of 15 (typescript:S3776), and this PR's role/aria attributes add to it. Two
changes, both moving code that has nothing to do with rendering a container:

- getPlacementStyle goes to module scope. It closes over nothing — only its
  own placement argument — so it was being rebuilt on every render for no
  reason, and its switch counted against the component.
- The two effects that only a modal survey WITHOUT an overlay needs, Escape
  handling and the open announcement, move into useNoOverlayModal. Between
  them they carried most of the component's branching: two guard clauses, two
  nested handlers and the modifier-key check. They share one enabled
  condition, which is now computed once at the call site. This follows the
  package's existing use-focus-trap / use-click-outside-hook pattern.

The component is left with one effect (click-outside) and the render.

No local rule measures cognitive complexity, so this deliberately reduces
with margin rather than shaving a point off; SonarQube on the PR is the
check. Behaviour is unchanged, verified against the built bundle on both
modal paths: with no overlay, aria-modal is absent, the live region gets "A
survey has opened at the end of the page." and Escape closes the survey; with
a dark overlay, aria-modal is "true", nothing is announced (the focus trap
moving focus is the announcement) and Escape still closes. Both keep the
single h1 and the named form landmark.
